[PRIVILEGES] :sparkles: Add user Privileges - Refactoring tests.
